refactor(architecture): FSD blocks/pages 레이어 구축 및 전체 아키텍처 완성
FSD 아키텍처 리팩토링의 마지막 단계로, `blocks` 레이어를 구축하고 `pages` 레이어를 최종 재구성하여 전체 아키텍처를 완성합니다.

기존 `components` 폴더에 남아있던 `Sidebar`, `Header` 등 공용 레이アウト 컴포넌트와 페이지별 뷰 컴포넌트들을 `blocks` 레이어로 이동 및 재구성했습니다. 특히, 반복되는 레이아웃을 `DashboardLayout`, `AuthLayout` 등 재사용 가능한 블록으로 추상화하여 코드 중복을 제거하고 일관성을 확보했습니다.

최종적으로 모든 `pages`는 FSD의 하위 레이어(`blocks`, `features`, `domain`)에서 만든 모듈들을 가져와 조립하는 단순하고 선언적인 역할만 수행하도록 리팩토링하여 '관심사의 완벽한 분리'를 달성했습니다.

**주요 변경 사항:**
- `src/blocks` 디렉토리 생성 및 `Sidebar`, `DashboardHeader`, `MainDashboardView` 등 21개의 UI 조합 블록 구조화
- 8개의 모든 `pages` 컴포넌트를 `blocks`와 `features`를 조립하는 형태로 전면 리팩토링
- `pages`에서 사용하던 `import` 경로를 FSD 구조에 맞게 최종 업데이트
- 잔여 `components` 폴더 정리 및 삭제 (figma 관련 코드 제외)

**이 커밋을 통해 프로젝트는 FSD의 4-Layer(common, domain, features, blocks) + pages 구조를 완전히 갖추게 되어, 유지보수성과 확장성이 극대화된 아키텍처로의 전환을 완료합니다.**
