+++ src/blocks/EnhancedPlatformCard/index.jsx
