+++ src/pages/Dashboard.jsx
@@ -1,36 +1,16 @@
-import { motion } from 'framer-motion';
-
-// 컴포넌트들 import
-import Sidebar from '../components/dashboard/sidebar.jsx';
-import DashboardHeader from '../components/dashboard/dashboard_header.jsx';
-import MainDashboardView from '../components/dashboard/main_dashboard_view.jsx';
+// FSD Blocks 조합
+import DashboardLayout from '../blocks/DashboardLayout';
+import MainDashboardView from '../blocks/MainDashboardView';
 
 /**
- * Dashboard 컴포넌트
- * 메인 대시보드 페이지 (MainDashboardView만 렌더링)
+ * Dashboard 페이지
+ * 메인 대시보드 - FSD 아키텍처의 조립 레이어
  */
 const Dashboard = () => {
-
   return (
-    <div className="h-screen bg-white dark:bg-gray-900 transition-colors duration-300 flex overflow-hidden">
-      {/* Subtle gradient background overlay */}
-      <div className="absolute inset-0 bg-gradient-to-br from-blue-50/30 via-purple-50/20 to-indigo-50/30 dark:from-blue-950/20 dark:via-purple-950/10 dark:to-indigo-950/20" />
-      
-      <Sidebar current_view="dashboard" />
-      
-      <div className="flex-1 flex flex-col relative z-10">
-        <DashboardHeader current_view="dashboard" />
-        <main className="flex-1 overflow-y-auto">
-          <motion.div
-            initial={{ opacity: 0, y: 20 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ duration: 0.3 }}
-          >
-            <MainDashboardView />
-          </motion.div>
-        </main>
-      </div>
-    </div>
+    <DashboardLayout currentView="dashboard" title="대시보드">
+      <MainDashboardView />
+    </DashboardLayout>
   );
 };
