+++ src/pages/LandingPage.jsx
@@ -1,15 +1,18 @@
 import { useState } from 'react';
 import { use_dark_mode } from '../common/hooks/use_dark_mode.js';
 import FloatingNav from '../common/ui/FloatingNav.jsx';
-import HeroSection from '../components/landing/HeroSection.jsx';
-import FeaturesSection from '../components/landing/FeaturesSection.jsx';
-import TransformSection from '../components/landing/TransformSection.jsx';
-import ReadySection from '../components/landing/ReadySection.jsx';
-import Footer from '../components/landing/Footer.jsx';
 import InfoModal from '../common/ui/InfoModal.jsx';
 
+// FSD Blocks 조합
+import HeroSection from '../blocks/HeroSection';
+import FeaturesSection from '../blocks/FeaturesSection';
+import TransformSection from '../blocks/TransformSection';
+import ReadySection from '../blocks/ReadySection';
+import Footer from '../blocks/Footer';
+
 /**
- * 랜딩 페이지 메인 컴포넌트
+ * LandingPage 페이지
+ * 랜딩 페이지 - FSD 아키텍처의 조립 레이어
  */
 export default function LandingPage() {
   // 다크모드 훅 사용
@@ -20,14 +23,13 @@ export default function LandingPage() {
   const [modal_title, set_modal_title] = useState('');
   const [modal_content, set_modal_content] = useState('');
 
-  // 모달 열기 핸들러
+  // 모달 핸들러
   const handle_open_modal = (title, content) => {
     set_modal_title(title);
     set_modal_content(content);
     set_modal_is_open(true);
   };
 
-  // 모달 닫기 핸들러
   const handle_close_modal = () => {
     set_modal_is_open(false);
     set_modal_title('');
@@ -36,10 +38,8 @@ export default function LandingPage() {
 
   return (
     <div className="relative overflow-hidden bg-white dark:bg-gray-900 transition-colors duration-300">
-      {/* Floating Navigation */}
       <FloatingNav />
       
-      {/* Content */}
       <div className="relative z-10">
         <HeroSection />
         <FeaturesSection />
@@ -48,7 +48,6 @@ export default function LandingPage() {
         <Footer on_open_modal={handle_open_modal} />
       </div>
 
-      {/* Info Modal */}
       <InfoModal
         isOpen={modal_is_open}
         onClose={handle_close_modal}
