+++ src/pages/SettingsPage.jsx
@@ -1,38 +1,14 @@
-import { motion } from 'framer-motion';
-
-// 컴포넌트들 import
-import Sidebar from '../components/dashboard/sidebar.jsx';
-import DashboardHeader from '../components/dashboard/dashboard_header.jsx';
-import SettingsView from '../components/dashboard/settings_view.jsx';
+import DashboardLayout from '../blocks/DashboardLayout';
+import SettingsView from '../blocks/SettingsView';
 
 /**
  * SettingsPage 컴포넌트
- * 환경설정 페이지
+ * 환경설정 페이지를 위한 레이아웃과 뷰 조립
  */
-const SettingsPage = () => {
-  
-
-  return (
-    <div className="h-screen bg-white dark:bg-gray-900 transition-colors duration-300 flex overflow-hidden">
-      {/* Subtle gradient background overlay */}
-      <div className="absolute inset-0 bg-gradient-to-br from-blue-50/30 via-purple-50/20 to-indigo-50/30 dark:from-blue-950/20 dark:via-purple-950/10 dark:to-indigo-950/20" />
-      
-      <Sidebar current_view="settings" />
-      
-      <div className="flex-1 flex flex-col relative z-10">
-        <DashboardHeader current_view="settings" />
-        <main className="flex-1 overflow-y-auto">
-          <motion.div
-            initial={{ opacity: 0, y: 20 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ duration: 0.3 }}
-          >
-            <SettingsView />
-          </motion.div>
-        </main>
-      </div>
-    </div>
-  );
-};
+const SettingsPage = () => (
+  <DashboardLayout currentView="settings" title="설정">
+    <SettingsView />
+  </DashboardLayout>
+);
 
 export default SettingsPage;
